chore: 拆分扩展方法
